Move input filters to separate component
